pyproject.toml
[project]
name = "finflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
]
